Dockerfile: update base image version
Changes:

- Update a base image of the last stage,
  docker:20.10.2-dind -> docker:20.10.12-dind.
  It is driven by a need to update ostree version (>= 2021.1).

- Remove sota-tools compilation and added garage-sign
  downloading and installation.
  Since we started to use fiotools, then garage-push/check/deploy
  became not needed, just garage-sign is required, which does not
  require any compilation, just download.

- Remove WIC utility installation.
  Since we started to use a loop device (losetup),
  then the WIC utility no longer needed.

- Install docker-compose by means of `apk` at the final stage.
  Installation of docker-compose by pip3 implies compilation of some of
  its dependencies. The compilation takes quite some time, and then
  the resultant artifacts have to be copied from the "build" stage to
  the final stage. Since we really don't need specific version of
  docker-compose, it just has to be higher than 1.27, it can be installed
  by `apk` without any compilation.

- Remove some obsolete stuff, ota-publish and ota-dockerapp.
  They looks like some leftovers from the previous life.

The following is history behind the given change.

Initially, I had a humble intention - update an ostree, from 2020.3 to
at least 2021.1, in order to support a new type of delta indexes.
There are two ways how it can be done, custom compilation or
installation by `apk`. Ostree compilation is a headache, and the author,
being a lazy person, decided to go with the `apk` approach.
To do it, the base image of the final build stage had to be updated,
from docker:20.10.2-dind to >=docker:20.10.8-dind.
The later installs python 3.9 while the base image of the build stage
installs python 3.8 and builds docker-compose based on it.
Therefore, the base image of the build stage had to be upgraded. This
led to another two problems:
1. For some reason the aktualizr compilation fails if conducted
   on hub.docker.io in this case (newer version of alpine).
   It's really odd that it works well on a local host.

2. Copying of the compiled artifacts from the build stage to the final
   stage doesn't do the job, docker-compose is not runnable
   on a resultant image (copy of site-packages and docker-compose binary).

The author, again being a lazy person, decided to overcome
the above stated issues, instead of solving them.
"Why do we need to compile aktualizr at all?" - thought struck the author,
we just need garage-sign which can be downloaded.

Another existential question is
"why installation of the docker-compose python package via pip requires
compilation of some of its dependencies?". It takes lots of time and after
that you need to figure out what exactly have to be copied from the build
stage to the final stage to make it work. In addition to that, the
base images of the build and final stages are different, what makes
another subtle difficulties.
Since we really don't need specific version of docker-compose, then
why not install it via `apk` (no compilation) taking into account
that we need to upgrade the base image anyway (ostree) and the
upgraded image installs docker-compose >= 1.27 (1.29).

Signed-off-by: Mike Sul <mike.sul@foundries.io>