dockerfile: Control layers manifest injection
Add an environment variable to control whether to add the layers
manifest into an app manifest or not. This allows to produce
OCI-compliant app manifest if `OCI_COMPLIANT_APP` is set to `"1"`.

Eventually, it will allows us controlling it through factory
configuration once corresponding changes to source.f.io are made.

Signed-off-by: Mike Sul <mike.sul@foundries.io>